% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesaverse.R
\docType{package}
\name{pesaverse-package}
\alias{pesaverse}
\alias{pesaverse-package}
\title{pesaverse: Easily Install and Load the Set of Packages}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A list of packages I use frequently, easily attaching in the 'pesaverse' package. This is a better option than editing the .Rprofile to guarantee reproducibility, moreover, this gives information about the list of attached packages and their versions. This is directly inspired by Hadley Wickham's 'tidyverse' package and borrows some of its code and look.
}
\author{
\strong{Maintainer}: Paul Efren Santos Andrade \email{paulefrens@gmail.com}

}
\keyword{internal}
