% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{pesaverse_deps}
\alias{pesaverse_deps}
\title{List all pesaverse dependencies}
\usage{
pesaverse_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
pesaverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\description{
List all pesaverse dependencies
}
