% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pesaverse_packages}
\alias{pesaverse_packages}
\title{List all packages in the pesaverse}
\usage{
pesaverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include pesaverse in the list?}
}
\description{
List all packages in the pesaverse
}
\examples{
pesaverse_packages()
}
